wonder_mail_3_mid is finally split correctly
