+++ ld_script.txt
@@ -160,6 +160,8 @@ SECTIONS {
         src/mailbox.o(.text);
         src/pelipper_board.o(.text);
         src/wonder_mail_3_mid.o(.text);
+        src/mission_reward.o(.text);
+        src/code_8072F3C.o(.text);
         asm/wonder_mail_3_mid.o(.text);
         src/wonder_mail_3_1.o(.text);
         src/makuhita_dojo1.o(.text);
@@ -500,6 +502,8 @@ SECTIONS {
         src/mailbox.o(.rodata);
         src/pelipper_board.o(.rodata);
         src/wonder_mail_3_mid.o(.rodata);
+        src/mission_reward.o(.rodata);
+        src/code_8072F3C.o(.rodata);
         src/makuhita_dojo1.o(.rodata);
         src/makuhita_dojo2.o(.rodata);
         src/wonder_mail_4.o(.rodata);
